Static methods and main distinguished
